Time delta is now a double

+++ src/main/java/AztecChallenge/Minigames/TempleAssault/TempleAssault.java
@@ -25,7 +25,7 @@ public class TempleAssault extends Engine {
     }
 
     @Override
-    protected void tick(long timedelta) {
+    protected void tick(double timedelta) {
